refactor: Report 관련 Enum 패키지 분리 - ReportType Enum 추가
