+++ src/main/java/com/dongyang/dongpo/domain/report/enums/ReportStatus.java
@@ -1,0 +1,7 @@
+package com.dongyang.dongpo.domain.report.enums;
+
+public enum ReportStatus {
+    OPEN,
+    PROCEEDING,
+    CLOSED
+}
