+++ src/main/java/com/dongyang/dongpo/domain/report/enums/ReportType.java
@@ -1,0 +1,6 @@
+package com.dongyang.dongpo.domain.report.enums;
+
+public enum ReportType {
+    REVIEW,
+    STORE
+}
